removing boilerplate code for widget items
