+++ src/Widgets/Menu/CommonItem.php
@@ -3,18 +3,13 @@
 namespace Tkui\Widgets\Menu;
 
 use Tkui\Options;
-use SplSubject;
-use Tkui\Support\WithObservable;
-use Tkui\TclTk\TclOptions;
+use Tkui\Widgets\Common\SubjectItem;
 
 /**
  * Base for a menu item.
  */
-abstract class CommonItem implements SplSubject
+abstract class CommonItem extends SubjectItem
 {
-    use WithObservable;
-
-    private Options $options;
     private int $id;
 
     // TODO: id generator ?
@@ -22,9 +17,8 @@ abstract class CommonItem implements SplSubject
 
     public function __construct(array|Options $options = [])
     {
+        parent::__construct($options);
         $this->id = self::generateId();
-        $this->options = $this->createOptions()
-                              ->with($options);
     }
 
     private static function generateId(): int
@@ -32,28 +26,7 @@ abstract class CommonItem implements SplSubject
         return ++self::$idIterator;
     }
 
-    protected function createOptions(): Options
-    {
-        return new TclOptions();
-    }
-
-    public function __set($name, $value)
-    {
-        $this->options->$name = $value;
-        $this->notify();
-    }
-
-    public function __get($name)
-    {
-        return $this->options->$name;
-    }
-
     abstract public function type(): string;
-
-    public function options(): Options
-    {
-        return $this->options;
-    }
 
     // TODO: identificable interface
     public function id(): int
