+++ src/Widgets/NotebookTab.php
@@ -2,10 +2,9 @@
 
 namespace Tkui\Widgets;
 
-use SplSubject;
 use Tkui\Options;
-use Tkui\Support\WithObservable;
 use Tkui\TclTk\TclOptions;
+use Tkui\Widgets\Common\SubjectItem;
 use Tkui\Widgets\Common\WithUnderlinedLabel;
 
 /**
@@ -16,24 +15,19 @@ use Tkui\Widgets\Common\WithUnderlinedLabel;
  * @property string $image TODO
  * @property string $compound TODO
  * @property int $underline
- *
- * @todo Just options instance ?
  */
-class NotebookTab implements SplSubject
+class NotebookTab extends SubjectItem
 {
     use WithUnderlinedLabel;
-    use WithObservable;
 
     private Widget $container;
-    private Options $options;
 
     public function __construct(Widget $container, string $title, array|Options $options = [])
     {
-        $optionsObj = $this->createOptions()->with($options);
-        $optionsObj->text = $this->removeUnderlineChar($title);
-        $optionsObj->underline = $this->detectUnderlineIndex($title);
-        $this->options = $this->createOptions()->with($optionsObj);
+        parent::__construct($options);
         $this->container = $container;
+        $this->text = $this->removeUnderlineChar($title);
+        $this->underline = $this->detectUnderlineIndex($title);
     }
 
     protected function createOptions(): Options
@@ -47,22 +41,6 @@ class NotebookTab implements SplSubject
             'compound' => null,
             'underline' => null,
         ]);
-    }
-
-    public function __get($name)
-    {
-        return $this->options->$name;
-    }
-
-    public function __set($name, $value)
-    {
-        $this->options->$name = $value;
-        $this->notify();
-    }
-
-    public function options(): Options
-    {
-        return $this->options;
     }
 
     public function container(): Widget
