+++ src/Widgets/TreeView/Column.php
@@ -6,6 +6,7 @@ namespace Tkui\Widgets\TreeView;
 
 use Tkui\Options;
 use Tkui\TclTk\TclOptions;
+use Tkui\Widgets\Common\SubjectItem;
 use Tkui\Widgets\Consts\Anchor;
 
 /**
@@ -15,16 +16,15 @@ use Tkui\Widgets\Consts\Anchor;
  * @property bool $stretch
  * @property int $width
  */
-class Column
+class Column extends SubjectItem
 {
-    private Options $options;
     private Header $header;
 
     final public function __construct(string $id, Header $header, array|Options $options = [])
     {
-        $options['id'] = $id;
+        parent::__construct($options);
+        $this->id = $id;
         $this->header = $header;
-        $this->options = $this->createOptions()->with($options);
     }
 
     protected function createOptions(): Options
@@ -38,35 +38,12 @@ class Column
         ]);
     }
 
-    public function options(): Options
-    {
-        return $this->options;
-    }
-
-    /**
-     * @param string $name
-     * @return mixed
-     */
-    public function __get($name)
-    {
-        return $this->options->$name;
-    }
-
-    /**
-     * @param string $name
-     * @param mixed $value
-     */
-    public function __set($name, $value)
-    {
-        $this->options->$name = $value;
-    }
-
     public function header(): Header
     {
         return $this->header;
     }
 
-    public static function create(string $id, string $header): self
+    public static function create(string $id, string $header): static
     {
         return new static($id, new Header($header));
     }
