+++ src/Widgets/TreeView/Header.php
@@ -7,6 +7,7 @@ namespace Tkui\Widgets\TreeView;
 use Tkui\Image;
 use Tkui\Options;
 use Tkui\TclTk\TclOptions;
+use Tkui\Widgets\Common\SubjectItem;
 use Tkui\Widgets\Consts\Anchor;
 
 /**
@@ -17,15 +18,12 @@ use Tkui\Widgets\Consts\Anchor;
  *
  * @todo Just extend from TclOptions ?
  */
-class Header
+class Header extends SubjectItem
 {
-    private Options $options;
-
     public function __construct(string $text, array|Options $options = [])
     {
-        $this->options = $this->createOptions()->with($options + [
-            'text' => $text,
-        ]);
+        parent::__construct($options);
+        $this->text = $text;
     }
 
     protected function createOptions(): Options
@@ -36,27 +34,5 @@ class Header
             'anchor' => null,
             'command' => null,
         ]);
-    }
-
-    public function options(): Options
-    {
-        return $this->options;
-    }
-    /**
-     * @param string $name
-     * @return mixed
-     */
-    public function __get($name)
-    {
-        return $this->options->$name;
-    }
-
-    /**
-     * @param string $name
-     * @param mixed $value
-     */
-    public function __set($name, $value)
-    {
-        $this->options->$name = $value;
     }
 }
